tests for reducers
