--- .babelrc
@@ -1,30 +1,0 @@
-{
-  "env": {
-    "presets": [
-      [
-        "es2015",
-        {
-          "modules": false
-        }
-      ],
-      "stage-2",
-      "latest",
-      "react",
-      [
-        "env",
-        {
-          "targets": {
-            "browsers": [
-              "last 2 versions"
-            ]
-          },
-          "useBuiltIns": false
-        }
-      ]
-    ],
-    "plugins": [
-      "react-hot-loader/babel",
-      "transform-object-rest-spread"
-    ]
-  }
-}
