+++ .gitignore
@@ -1,5 +1,4 @@
 node_modules
 .idea
 .nyc_output
-public/js
-public/css
+public/*.bundle.js
