+++ public/admin/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Chat App</title>
+    <title>Notifier Administrator</title>
 </head>
 <body style="background:#1b6afd;margin:0;padding:0;">
 <div id="app"></div>
