+++ public/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Chat App</title>
+    <title>Notifier Prototype</title>
 </head>
 <body style="background:#0043a2;margin:0;padding:0;">
 <div id="app"></div>
